+++ app/infrastructure/kafka/kafka_message.py
@@ -11,5 +11,5 @@ class KafkaMessage:
     scale_type: str
     duration: int
     bpm: int
-    figure: int
+    figure: float
     octaves: int
